+++ spring-core/dependency-injection/src/main/java/dangminhphuc/dev/demo/javabased/User.java
@@ -1,4 +1,4 @@
-package dangminhphuc.dev.demo.annotation;
+package dangminhphuc.dev.demo.javabased;
 
 public class User {
     private final String name;
